requirements.txt
apify-client
html2text
pandas
tqdm
playwright
pydantic
python-dotenv